[issue-122] Moved the distance matrix directory to commutes. It should belong there
